테스트 액셀, 레코드 및 데이터 추가

+++ TestRecord/Excel2Records.cs
@@ -1,7 +1,8 @@
 namespace Excel2;
 
+// Excel1의 SameNameSheet와 이름이 같지만, namespace가 다르기 때문에 충돌이 발생하지 않음
 [StaticDataRecord("Excel2", "SameNameSheet")]
-public sealed record SameNameSheet(string Name, int Score);
+public sealed record SameNameSheet(string Name, float Score);
 
 public sealed record SubjectData([Key][ColumnName("Subject")] string Name, int Score);
